move seed into common and add SpType
